Researcher in machine learning and healthcare
Passionate and motivated OR researcher working on machine learning applications for healthcare, with a strong focus on breast
cancer. Proactive and results-driven, I actively contribute to the field with two years of hands-on experience. With an unwavering positivity and commitment, I strive to advance scientific knowledge and make a meaningful difference in people's lives.